First round of fixes to tests
