+++ st2actions/tests/unit/policies/test_concurrency.py
@@ -24,10 +24,12 @@ from st2common.models.db.action import LiveActionDB
 from st2common.persistence.action import LiveAction
 from st2common.persistence.policy import Policy
 from st2common.services import action as action_service
+from st2common.transport.execution import ActionExecutionPublisher
 from st2common.transport.liveaction import LiveActionPublisher
 from st2common.transport.publishers import CUDPublisher
 from st2tests import DbTestCase, EventletTestCase
 from st2tests.fixturesloader import FixturesLoader
+from st2tests.mocks.execution import MockExecutionPublisher, MockExecutionPublisherNonBlocking
 from st2tests.mocks.liveaction import MockLiveActionPublisher, MockLiveActionPublisherNonBlocking
 from st2tests.mocks import runner
 
@@ -59,7 +61,7 @@ SCHEDULED_STATES = [
             mock.MagicMock(return_value=runner))
 @mock.patch.object(
     CUDPublisher, 'publish_update',
-    mock.MagicMock(side_effect=MockLiveActionPublisher.publish_update))
+    mock.MagicMock(side_effect=MockExecutionPublisher.publish_update))
 @mock.patch.object(
     CUDPublisher, 'publish_create',
     mock.MagicMock(return_value=None))
@@ -172,7 +174,7 @@ class ConcurrencyPolicyTest(EventletTestCase, DbTestCase):
     # running in the same process.
     @mock.patch.object(
         LiveActionPublisher, 'publish_update',
-        mock.MagicMock(side_effect=MockLiveActionPublisherNonBlocking.publish_update))
+        mock.MagicMock(side_effect=MockExecutionPublisherNonBlocking.publish_update))
     def test_over_threshold_cancel_executions(self):
         policy_db = Policy.get_by_ref('wolfpack.action-2.concurrency.cancel')
         self.assertEqual(policy_db.parameters['action'], 'cancel')
@@ -237,6 +239,9 @@ class ConcurrencyPolicyTest(EventletTestCase, DbTestCase):
     @mock.patch.object(
         LiveActionPublisher, 'publish_state',
         mock.MagicMock(side_effect=MockLiveActionPublisherNonBlocking.publish_state))
+    @mock.patch.object(
+        ActionExecutionPublisher, 'publish_update',
+        mock.MagicMock(side_effect=MockExecutionPublisherNonBlocking.publish_update))
     def test_on_cancellation(self):
         policy_db = Policy.get_by_ref('wolfpack.action-1.concurrency')
         self.assertGreater(policy_db.parameters['threshold'], 0)
